refactor(ses): introduce makeEvaluate
